order_service/requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.28
requests==2.26.0
